feat: 결제 테이블 추가

+++ src/main/java/roomescape/service/payment/dto/PaymentConfirmInput.java
@@ -1,4 +1,4 @@
 package roomescape.service.payment.dto;
 
-public record PaymentConfirmInput(String orderId, Integer amount, String paymentKey) {
+public record PaymentConfirmInput(String orderId, Long amount, String paymentKey) {
 }
